+++ packages/docs/docs/.vuepress/components/Close.vue
@@ -1,0 +1,14 @@
+<template>
+  <svg
+    class="status-close-icon"
+    xmlns="http://www.w3.org/2000/svg"
+    width="32"
+    height="32"
+    viewBox="-4 -4 32 32"
+  >
+    <path
+      fill="currentColor"
+      d="m12 10.586l4.95-4.95l1.414 1.414l-4.95 4.95l4.95 4.95l-1.414 1.414l-4.95-4.95l-4.95 4.95l-1.414-1.414l4.95-4.95l-4.95-4.95L7.05 5.636z"
+    />
+  </svg>
+</template>