+++ packages/vuepress-theme-vt/enhanceApp.js
@@ -1,3 +1,5 @@
+import { $status, STATUS_HIDDEN_EVENT } from "./lib/status";
+
 export default ({
   Vue, // the version of Vue being used in the VuePress app
   options, // the options for the root Vue instance
@@ -6,6 +8,10 @@ export default ({
 }) => {
   Vue.mixin({
     methods: {
+      $closeCurrentStatus() {
+        $status.closeCurrentStatus();
+        this.$root.$emit(STATUS_HIDDEN_EVENT);
+      },
       $withLocale: function (path) {
         if (path.startsWith("/")) {
           return this.$localePath + path.replace(/^\//, "");